+++ src/index.js
@@ -1,7 +1,4 @@
-//* Imports
 const Server = require('./structures/server.js');
 const config = require('./config.json');
 
-//* Initialize the server
-const api = new Server(config);
-api.run();
+new Server(config).run();