+++ src/routers/index.js
@@ -1,8 +1,10 @@
+const package = require('../../package.json');
+
 module.exports = ({ server, config }) => {
     server.get('/', async () => {
         return {
-            statusCode: 200,
-            message: `API docs: ${config.docs}`
+            version: package.version,
+            message: `Hello World! API docs: ${config.docs}`
         };
     });
 };