+++ src/structures/server.js
@@ -5,34 +5,31 @@ const config = require('../config.json');
 
 const log = new Logger({
 	name: config.logs.name,
-	logToFile: config.logs.logToFile,
+    logToFile: config.logs.logToFile,
+    format: '{method} {status-colour}{status} &7{path} {time-colour}({time})'
 });
 
-const db = require('better-sqlite3')(config.database);
+const db = require('better-sqlite3')('mue.db');
 db.pragma('journal_mode = WAL'); // This makes sqlite FAST
-
-const GhostContentAPI = require('@tryghost/content-api'); // for getUpdate endpoint
-const ghost = new GhostContentAPI({
-	url: config.update.blogurl,
-	key: config.update.ghostapikey,
-	version: config.update.ghostversion
-});
+db.pragma('synchronous = OFF'); // disable sync, we only do read requests so hopefully corruption shouldn't occur
 
 module.exports = class Server {
     constructor(config) {
-        this.server = fastify();
+        this.server = fastify({
+            ignoreTrailingSlash: true
+        });
         this.config = config;
         this.log = log;
         this.db = db;
-        this.ghost = ghost;
     }
 
     addMiddleware() {
         this.server.register(log.fastify); // logger plugin
-        this.server.register(require('fastify-cors'));
+        if (!this.config.nginx) this.server.register(require('fastify-cors'));
         this.server.register(require('fastify-rate-limit'), {
             max: config.ratelimit.max,
-            timeWindow: config.ratelimit.timewin
+            timeWindow: config.ratelimit.timewin,
+            keyGenerator: (req) => { return req.headers['x-real-ip'] || req.raw.ip } // support for nginx
         });
     }
 
@@ -43,7 +40,7 @@ module.exports = class Server {
             if (error) this.log.error(`Unable to build routes:\n${error}`);
             files.forEach(file => {
                 require(`../routers/${file}`)(this);
-                this.log.info(`Built route in "${file}"`);
+                this.log.info(`Built ${file} route`);
             });
         });
     }
@@ -54,7 +51,7 @@ module.exports = class Server {
 
         await new Promise(resolve => setTimeout(resolve, 1000));
         this.server.listen(this.config.port, (error, address) =>
-            error ? this.log.error(`Unable to build the fastify instance:\n${error}`) : this.log.info(`Running the API server at ${address}`)
+            error ? this.log.error(`Unable to build the Fastify instance:\n${error}`) : this.log.info(`Running the API server at ${address}`)
         );
     }
 };